requirements.txt
ipykernel
ipython
jsonschema
nltk
numpy
networkx
pandas
scikit-learn
scikit-multilearn
scipy
texttable
torch
torchaudio
torchtext
torchvision
tqdm
gensim
flake8
python-dotenv
libpecos